requirements.txt
flask==3.0.3
flask-mail==0.10.0
python-dotenv==1.0.1


